ai_cheques/signature_extraction/requirements.txt
torch
ultralytics
clip
Pillow
numpy